+++ src/components/form/FormFields.tsx
@@ -1,10 +1,11 @@
 import React, { useRef, useState } from "react";
 
 import classnames from "../../classnames";
+import useAutoFocus from "../../utils/useAutoFocus";
+
 import FormError from "./FormError";
 import { useForm } from "./Form";
 import { FormFieldError } from "./typings";
-import useAutoFocus from "./useAutoFocus";
 import useDisabled from "./useDisabled";
 
 type HijackedProps = "className" | "name" | "onChange" | "required" | "value";
@@ -34,7 +35,7 @@ const makeFormField = (type: string) => {
     const inputRef = useRef<HTMLInputElement | null>(null);
     const [touched, setTouched] = useState<boolean>(false);
 
-    useAutoFocus(autoFocus || false, inputRef);
+    useAutoFocus(autoFocus, inputRef);
     useDisabled(name, disabled);
 
     const onBlur = () => setTouched(true);