+++ CHANGELOG.md
@@ -18,10 +18,12 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Support the `disabled` prop on the `RadioField` component.
 - Support the `disabled` prop on the `TextField` component.
 - Support the `disabled` prop on the `SelectField` component.
+- The `Dropdown` component.
 
 ### Changed
 
 - The value of the `SelectField` checkmark in the CSS from `✓` to `\2713` to make it more cross-browser compliant.
+- The `Button` component is now a `forwardRef`.
 
 ## [16.3.0] - 2020-01-07