+++ CHANGELOG.md
@@ -19,6 +19,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Support the `disabled` prop on the `TextField` component.
 - Support the `disabled` prop on the `SelectField` component.
 
+### Changed
+
+- The value of the `SelectField` checkmark in the CSS from `✓` to `\2713` to make it more cross-browser compliant.
+
 ## [16.3.0] - 2020-01-07
 
 ### Added