+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ## [Unreleased]
 
+## [16.4.0] - 2020-01-08
+
 ### Added
 
 - Support the `disabled` prop on the `Checkmark` component.
@@ -1117,7 +1119,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Initial checkin.
 - The `Warning` component.
 
-[unreleased]: https://github.com/CultureHQ/components/compare/v16.3.0...HEAD
+[unreleased]: https://github.com/CultureHQ/components/compare/v16.4.0...HEAD
+[16.4.0]: https://github.com/CultureHQ/components/compare/v16.3.0...v16.4.0
 [16.3.0]: https://github.com/CultureHQ/components/compare/v16.2.1...v16.3.0
 [16.2.1]: https://github.com/CultureHQ/components/compare/v16.2.0...v16.2.1
 [16.2.0]: https://github.com/CultureHQ/components/compare/v16.1.1...v16.2.0